pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "correlator-ge"
version = "0.0.1"
description = "Great Expectations plugin that emits OpenLineage events for automated incident correlation"
readme = "README.md"
authors = [
    { name = "Emmanuel King Kasulani", email = "kasulani@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring",
]
keywords = [
    "great-expectations",
    "openlineage",
    "data-quality",
    "observability",
    "lineage",
    "correlation",
    "incident-management",
]
requires-python = ">=3.9"
dependencies = [
    "great_expectations>=1.3.0",
    "openlineage-python>=1.0.0",
    "click>=8.1.0",
    "requests>=2.28.0",
    "PyYAML>=6.0",
]

[project.license]
text = "Apache-2.0"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.8.0",
    "mypy>=1.9.0",
    "isort>=5.13.0",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.7.0",
    "types-click>=7.1.0",
    "responses>=0.24.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/correlator-io/correlator-ge"
Repository = "https://github.com/correlator-io/correlator-ge"
Documentation = "https://github.com/correlator-io/correlator-ge#readme"
"Bug Tracker" = "https://github.com/correlator-io/correlator-ge/issues"
Correlator = "https://github.com/correlator-io/correlator"

[project.scripts]
ge-correlator = "ge_correlator.cli:cli"

[tool.hatch.build.targets.wheel]
packages = [
    "src/ge_correlator",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 88
target-version = [
    "py39",
    "py310",
    "py311",
    "py312",
    "py313",
]
include = "\\.pyi?$"
exclude = "/(\n    \\.eggs\n    | \\.git\n    | \\.hg\n    | \\.mypy_cache\n    | \\.tox\n    | \\.venv\n    | _build\n    | buck-out\n    | build\n    | dist\n    | venv\n    | tests/fixtures\n)/\n"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = [
    "src",
    "tests",
]

[tool.ruff]
line-length = 88
target-version = "py39"
src = [
    "src",
    "tests",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "ARG",
    "SIM",
    "PT",
    "RET",
    "PL",
    "RUF",
]
ignore = [
    "E501",
    "PLR0913",
    "PLR2004",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",
    "ARG",
    "PLR2004",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
overrides = [
    { module = "tests.*", disallow_untyped_defs = false },
    { module = "openlineage.*", ignore_missing_imports = true },
    { module = "great_expectations.*", ignore_missing_imports = true },
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = [
    "tests",
]
norecursedirs = [
    "*.egg",
    ".eggs",
    "dist",
    "build",
    "docs",
    ".tox",
    ".git",
    "__pycache__",
]
python_files = "test_*.py"
python_classes = "Test[!R]*"
python_functions = "test_*"
addopts = [
    "--cov=ge_correlator",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-branch",
    "--strict-markers",
    "-v",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = [
    "src/ge_correlator",
]
branch = true
omit = [
    "*/tests/*",
    "*/site-packages/*",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/interfaces.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pass",
    "raise NotImplementedError",
    "@abstractmethod",
    "def __repr__",
    "def __str__",
    "pragma: no cover",
    "def __debug__",
    "if self\\.debug",
    "raise AssertionError",
    "except ImportError",
    "except Exception as e:",
    "warn\\(",
    "if sys\\.platform ==",
    "if platform\\.",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.bandit]
exclude_dirs = [
    "tests",
    "docs",
]
skips = [
    "B101",
    "B404",
    "B603",
    "B607",
]

[dependency-groups]
dev = [
    "great-expectations>=1.3.0",
    "responses>=0.24.0",
    "types-attrs>=19.1.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]
